Fix compiler and linker warnings
- Add type casts where C++20 deprecates implicit enum/float/int
  conversions (rays color_space, rays-video video, reflex shape/style)
- Cast NSImageRep* to NSBitmapImageRep* in rays bitmap.mm loader
- Remove unused const get_data overload for WallShape in reflex
- Wrap not-implemented stubs (SoundPlayer::rewind, vibrate,
  Timers::find_timers) with #pragma to suppress -Wmissing-noreturn
- xot extconf: only add -lstdc++ on gcc; clang++ auto-links libc++
  and the explicit -lc++ produced an ld duplicate-library warning


[[7A053F39-54F2-4D0A-AF3F-814E0C8DE517]]

+++ ext/rays-video/video.cpp
@@ -28,7 +28,7 @@ RUCY_DEF4(initialize, width, height, fps, pixel_density)
 	*THIS = Rays::Video(
 		to<int>(width),
 		to<int>(height),
-		fps_ > 0 ? fps_ : Rays::Video::DEFAULT_FPS,
+		fps_ > 0 ? fps_ : (float) Rays::Video::DEFAULT_FPS,
 		to<float>(pixel_density));
 }
 RUCY_END
